+++ cmn-data-svc-client/src/main/java/org/acumos/cds/domain/MLPComment.java
@@ -1,0 +1,164 @@
+/*-
+ * ===============LICENSE_START=======================================================
+ * Acumos
+ * ===================================================================================
+ * Copyright (C) 2017 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+ * ===================================================================================
+ * This Acumos software file is distributed by AT&T and Tech Mahindra
+ * under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *  
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ * This file is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ===============LICENSE_END=========================================================
+ */
+
+package org.acumos.cds.domain;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.validation.constraints.Size;
+
+import org.hibernate.annotations.GenericGenerator;
+
+/**
+ * A comment in a thread.
+ */
+@Entity
+@Table(name = "C_COMMENT")
+public class MLPComment extends MLPTimestampedEntity implements Serializable {
+
+	private static final long serialVersionUID = 4666733550951208705L;
+
+	@Id
+	@GeneratedValue(generator = "customUseOrGenerate")
+	@GenericGenerator(name = "customUseOrGenerate", strategy = "org.acumos.cds.util.UseExistingOrNewUUIDGenerator")
+	@Column(name = "COMMENT_ID", updatable = false, nullable = false, columnDefinition = "CHAR(36)")
+	@Size(max = 36)
+	private String commentId;
+
+	@Column(name = "THREAD_ID", nullable = false, columnDefinition = "CHAR(36)")
+	@Size(max = 36)
+	private String threadId;
+
+	@Column(name = "PARENT_ID", columnDefinition = "CHAR(36)")
+	@Size(max = 36)
+	private String parentId;
+
+	@Column(name = "USER_ID", nullable = false, columnDefinition = "CHAR(36)")
+	@Size(max = 36)
+	private String userId;
+
+	@Column(name = "TEXT", nullable = false, columnDefinition = "VARCHAR(8192)")
+	@Size(max = 8192)
+	private String text;
+
+	@Column(name = "URL", columnDefinition = "VARCHAR(512)")
+	@Size(max = 512)
+	private String url;
+
+	/**
+	 * No-arg constructor
+	 */
+	public MLPComment() {
+		// no-arg constructor
+	}
+
+	/**
+	 * This constructor accepts the required fields; i.e., the minimum that the user
+	 * must supply to create a valid instance.
+	 * 
+	 * @param threadId
+	 *            Thread ID
+	 * @param userId
+	 *            User ID of author
+	 * @param text
+	 *            Comment text
+	 */
+	public MLPComment(String threadId, String userId, String text) {
+		this.threadId = threadId;
+		this.userId = userId;
+		this.text = text;
+	}
+
+	public String getCommentId() {
+		return commentId;
+	}
+
+	public void setCommentId(String commentId) {
+		this.commentId = commentId;
+	}
+
+	public String getThreadId() {
+		return threadId;
+	}
+
+	public void setThreadId(String threadId) {
+		this.threadId = threadId;
+	}
+
+	public String getParentId() {
+		return parentId;
+	}
+
+	public void setParentId(String parentId) {
+		this.parentId = parentId;
+	}
+
+	public String getUserId() {
+		return userId;
+	}
+
+	public void setUserId(String userId) {
+		this.userId = userId;
+	}
+
+	public String getText() {
+		return text;
+	}
+
+	public void setText(String text) {
+		this.text = text;
+	}
+
+	public String getUrl() {
+		return url;
+	}
+
+	public void setUrl(String url) {
+		this.url = url;
+	}
+
+	@Override
+	public boolean equals(Object that) {
+		if (that == null)
+			return false;
+		if (!(that instanceof MLPComment))
+			return false;
+		MLPComment thatObj = (MLPComment) that;
+		return Objects.equals(commentId, thatObj.commentId);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(commentId, threadId, userId, text, url);
+	}
+
+	@Override
+	public String toString() {
+		return this.getClass().getName() + "[commentId=" + commentId + ", threadId=" + threadId + ", userId=" + userId
+				+ ", text=" + text + ", url=" + url + "]";
+	}
+
+}