+++ cmn-data-svc-server/src/main/java/org/acumos/cds/specification/MLPSolutionSpecification.java
@@ -21,6 +21,12 @@ public class MLPSolutionSpecification implements Specification<MLPSolution> {
 
 	private SearchCriterion criterion;
 
+	/**
+	 * Creates a search specification with the specified criterion.
+	 * 
+	 * @param criteria
+	 *            SearchCriterion
+	 */
 	public MLPSolutionSpecification(final SearchCriterion criteria) {
 		this.criterion = criteria;
 	}
@@ -37,7 +43,8 @@ public class MLPSolutionSpecification implements Specification<MLPSolution> {
 			return builder.greaterThanOrEqualTo(queryRoot.<String>get(criterion.getKey()),
 					criterion.getValue().toString());
 		} else if (SearchOperation.LTE == criterion.getOperation()) {
-			return builder.lessThanOrEqualTo(queryRoot.<String>get(criterion.getKey()), criterion.getValue().toString());
+			return builder.lessThanOrEqualTo(queryRoot.<String>get(criterion.getKey()),
+					criterion.getValue().toString());
 		} else if (SearchOperation.IN == criterion.getOperation()) {
 			Expression<String> exp = queryRoot.<String>get(criterion.getKey());
 			if (criterion.getValue().getClass().isArray()) {